reorganize into submodules, add basic lexicon type
